Restructured test helper location and folder naming.
