+++ src/NUnitFramework/tests/ExecutionHooks/HookIdentifiers.cs
@@ -1,6 +1,6 @@
 // Copyright (c) Charlie Poole, Rob Prouse and Contributors. MIT License - see LICENSE.txt
 
-namespace NUnit.Framework.Tests.ExecutionHooks.Common
+namespace NUnit.Framework.Tests.ExecutionHooks
 {
     internal static class HookIdentifiers
     {
