+++ src/NUnitFramework/tests/ExecutionHooks/TestAttributes/ActivateAllSynchronousTestHooksAttribute.cs
